+++ src/components/b_modal.rs
@@ -7,17 +7,26 @@ pub fn BModal(
     #[prop(default = true)] has_background: bool,
     #[prop(default = true)] has_close_button: bool,
     #[prop(default = true)] is_dismissible: bool,
+    #[prop(optional, into)] on_close: Option<Callback<MouseEvent>>,
     is_active: RwSignal<bool>,
 ) -> impl IntoView {
+    let close_modal = move |event| {
+        is_active.set(false);
+
+        if let Some(oc) = on_close {
+            oc.call(event);
+        }
+    };
+
     view! {
         <Show when=move || is_active.get()>
             <div class="modal is-active">
                 <Show when=move || has_background>
                     <div
                         class="modal-background"
-                        on:click=move |_| {
+                        on:click=move |event| {
                             if is_dismissible {
-                                is_active.set(false)
+                                close_modal(event);
                             }
                         }
                     >
@@ -26,7 +35,7 @@ pub fn BModal(
                 {children()}
 
                 <Show when=move || has_close_button>
-                    <BModalClose on_click=move |_| is_active.set(false)/>
+                    <BModalClose on:click=close_modal/>
                 </Show>
             </div>
         </Show>
@@ -34,11 +43,8 @@ pub fn BModal(
 }
 
 #[component]
-pub fn BModalClose<F>(on_click: F) -> impl IntoView
-where
-    F: Fn(MouseEvent) + 'static,
-{
-    view! { <a class="modal-close is-large" on:click=on_click></a> }
+pub fn BModalClose() -> impl IntoView {
+    view! { <a class="modal-close is-large"></a> }
 }
 
 #[component]