+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+[Unreleased]
+
+# Added
+
+- Optional callback when modal is closed.
+
 ## [0.0.0-alpha.8] - 2024-04-08
 
 ### Added